requirements.txt
transformers
accelerate
runpod
pillow
opencv-python-headless
numpy
requests

